prefer different folder organization for install remove legacy visual studio solution move static variable NOINIT (in carve.hpp) into anonymous namespace
quoting from (https://stackoverflow.com/questions/357564/uses-for-anonymous-namespaces-in-header-files):
"It's a trick to allow a header only library to provide global objects, which would normally cause link errors. Putting an object in an anonymous namespace forces each compilation unit to create its own instance. The multiple instances don't matter in the case of _1, _2 etc. because they are stateless. Another way to do achieve the same effect is to use a static member of a template class, but then the client would have to type something like Dummy<>::_1 rather than just _1"